refactor(controller): split collection item helpers module
