+++ src/egui_app/controller/collection_items_helpers/io.rs
@@ -1,0 +1,71 @@
+use super::super::*;
+use hound::SampleFormat;
+use std::path::Path;
+
+pub(super) fn read_samples_for_normalization(
+    path: &Path,
+) -> Result<(Vec<f32>, hound::WavSpec), String> {
+    let bytes = crate::wav_sanitize::read_sanitized_wav_bytes(path)?;
+    let mut reader = hound::WavReader::new(std::io::Cursor::new(bytes))
+        .map_err(|err| format!("Invalid wav: {err}"))?;
+    let spec = reader.spec();
+    let samples = match spec.sample_format {
+        SampleFormat::Float => reader
+            .samples::<f32>()
+            .map(|s| s.map_err(|err| format!("Sample error: {err}")))
+            .collect::<Result<Vec<_>, _>>()?,
+        SampleFormat::Int => {
+            let scale = (1i64 << spec.bits_per_sample.saturating_sub(1)).max(1) as f32;
+            reader
+                .samples::<i32>()
+                .map(|s| {
+                    s.map(|value| value as f32 / scale)
+                        .map_err(|err| format!("Sample error: {err}"))
+                })
+                .collect::<Result<Vec<_>, _>>()?
+        }
+    };
+    Ok((samples, spec))
+}
+
+pub(super) fn write_normalized_wav(
+    path: &Path,
+    samples: &[f32],
+    spec: hound::WavSpec,
+) -> Result<(), String> {
+    let mut writer = hound::WavWriter::create(path, spec)
+        .map_err(|err| format!("Failed to write wav: {err}"))?;
+    for sample in samples {
+        writer
+            .write_sample(*sample)
+            .map_err(|err| format!("Failed to write sample: {err}"))?;
+    }
+    writer
+        .finalize()
+        .map_err(|err| format!("Failed to finalize wav: {err}"))
+}
+
+pub(super) fn replace_entry(list: &mut Vec<WavEntry>, old_path: &Path, new_entry: &WavEntry) {
+    if let Some(pos) = list
+        .iter()
+        .position(|entry| entry.relative_path == old_path)
+    {
+        list[pos] = new_entry.clone();
+    } else {
+        list.push(new_entry.clone());
+    }
+    list.sort_by(|a, b| a.relative_path.cmp(&b.relative_path));
+}
+
+pub(super) fn file_metadata(path: &Path) -> Result<(u64, i64), String> {
+    let metadata = std::fs::metadata(path)
+        .map_err(|err| format!("Failed to read {}: {err}", path.display()))?;
+    let modified_ns = metadata
+        .modified()
+        .map_err(|err| format!("Missing modified time for {}: {err}", path.display()))?
+        .duration_since(std::time::SystemTime::UNIX_EPOCH)
+        .map_err(|_| "File modified time is before epoch".to_string())?
+        .as_nanos() as i64;
+    Ok((metadata.len(), modified_ns))
+}
+
